Update
Level handling is mostly done except for the game over condition that I can't figure out how to trigger. I believe the midgame spawning mechanics are completely finished. The score handling is not fully implemented here, as I realizes it is mostly the responsibility of the actors more so than the student world. addToScore() function added to student world for actors to access and send amounts to. Water squirt has been updated to spawn at the correct location, however, it looks slightly different from the demo, not sure.